message_browser/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pymysql==1.1.1
python-dotenv==1.0.1
